Cleanup package names
