+++ accordionrecycler/src/main/java/com/izikode/izilib/accordionrecycler/AccordionRecyclerPosition.kt
@@ -1,4 +1,4 @@
-package com.izikode.accordionrecycler
+package com.izikode.izilib.accordionrecycler
 
 enum class AccordionRecyclerPosition {
