--- app/src/main/java/com/izikode/accordionrecyclerdemo/data/GrayData.kt
@@ -1,9 +1,0 @@
-package com.izikode.accordionrecyclerdemo.data
-
-import com.izikode.accordionrecyclerdemo.viewholder.GrayViewHolder
-
-class GrayData(
-
-    override var text: String
-
-) : ColorData(GrayViewHolder.VIEW_TYPE, arrayOf())
