--- app/src/main/java/com/izikode/accordionrecyclerdemo/data/PinkData.kt
@@ -1,9 +1,0 @@
-package com.izikode.accordionrecyclerdemo.data
-
-import com.izikode.accordionrecyclerdemo.viewholder.PinkViewHolder
-
-class PinkData(
-
-        override var text: String
-
-) : ColorData(PinkViewHolder.VIEW_TYPE, arrayOf())
