--- app/src/main/java/com/izikode/accordionrecyclerdemo/data/RedData.kt
@@ -1,9 +1,0 @@
-package com.izikode.accordionrecyclerdemo.data
-
-import com.izikode.accordionrecyclerdemo.viewholder.RedViewHolder
-
-class RedData(
-
-        override var text: String
-
-) : ColorData(RedViewHolder.VIEW_TYPE, arrayOf())
