--- app/src/main/java/com/izikode/accordionrecyclerdemo/data/WhiteData.kt
@@ -1,9 +1,0 @@
-package com.izikode.accordionrecyclerdemo.data
-
-import com.izikode.accordionrecyclerdemo.viewholder.WhiteViewHolder
-
-class WhiteData(
-
-        override var text: String
-
-) : ColorData(WhiteViewHolder.VIEW_TYPE, arrayOf())
