+++ app/src/main/java/com/izikode/izilib/accordionrecyclerdemo/data/GrayData.kt
@@ -1,0 +1,9 @@
+package com.izikode.izilib.accordionrecyclerdemo.data
+
+import com.izikode.izilib.accordionrecyclerdemo.viewholder.GrayViewHolder
+
+class GrayData(
+
+    override var text: String
+
+) : ColorData(GrayViewHolder.VIEW_TYPE, arrayOf())
