+++ app/src/main/java/com/izikode/izilib/accordionrecyclerdemo/data/PinkData.kt
@@ -1,0 +1,9 @@
+package com.izikode.izilib.accordionrecyclerdemo.data
+
+import com.izikode.izilib.accordionrecyclerdemo.viewholder.PinkViewHolder
+
+class PinkData(
+
+        override var text: String
+
+) : ColorData(PinkViewHolder.VIEW_TYPE, arrayOf())
