+++ app/src/main/java/com/izikode/izilib/accordionrecyclerdemo/data/RedData.kt
@@ -1,0 +1,9 @@
+package com.izikode.izilib.accordionrecyclerdemo.data
+
+import com.izikode.izilib.accordionrecyclerdemo.viewholder.RedViewHolder
+
+class RedData(
+
+        override var text: String
+
+) : ColorData(RedViewHolder.VIEW_TYPE, arrayOf())
