+++ app/src/main/java/com/izikode/izilib/accordionrecyclerdemo/data/WhiteData.kt
@@ -1,0 +1,9 @@
+package com.izikode.izilib.accordionrecyclerdemo.data
+
+import com.izikode.izilib.accordionrecyclerdemo.viewholder.WhiteViewHolder
+
+class WhiteData(
+
+        override var text: String
+
+) : ColorData(WhiteViewHolder.VIEW_TYPE, arrayOf())
